+++ projects/pastanaga-angular/src/lib/chips/chips.module.ts
@@ -3,10 +3,13 @@ import { CommonModule } from '@angular/common';
 import { ChipCloseableComponent } from './chip-closeable/chip-closeable.component';
 import { PaAvatarModule } from '../avatar/avatar.module';
 import { PaButtonModule } from '../button/button.module';
+import { ChipSelectionableComponent } from './chip-selectionable/chip-selectionable.component';
+import { PaIconModule } from '../icon/icon.module';
+import { PaFocusableModule } from '../focusable/focusable.module';
 
 @NgModule({
-    declarations: [ChipCloseableComponent],
-    imports: [CommonModule, PaAvatarModule, PaButtonModule],
-    exports: [ChipCloseableComponent],
+    imports: [CommonModule, PaAvatarModule, PaButtonModule, PaIconModule, PaFocusableModule],
+    declarations: [ChipCloseableComponent, ChipSelectionableComponent],
+    exports: [ChipCloseableComponent, ChipSelectionableComponent],
 })
 export class PaChipsModule {}